Updated Interface layer. Now it looks for connections.
TO-DO (didn't have enought time):
Make the suggestions clickable (so the user can tap on them instead of pressing "Search".
Implement a Loop that allows the information to refresh each 60 seconds (can be done with a "android.os.Handler" and a "java.lang.Runnable". How to fit it in the design?)
Improve the user interface. It works, but it can be improved.
Handle the exceptions (It is reasonable that we will have, at least, IO exceptions when the user don't have internet connection).
Refactor code. At least the code in the rest adapters feels a little bit messy. In general, almost al the code have been written with hurry (Except maybe the Application layer).
Test the code. It should be tested before it's written. The lack of permission to use external libraries (jUnit) and the time pressure are the reasons why it haven't been done.

Bugfixes:
URL for downloading connections (ConnectionsRestAdapter) had a typo. Fixed.